编译二进制文件；修复编译问题

+++ geodesic_algorithm_exact.h
@@ -135,8 +135,8 @@ private:
 	double start[6];
 	interval_pointer i_new[5];
 
-	long unsigned m_queue_max_size;			//used for statistics
-	unsigned m_iterations;			//used for statistics
+	size_t m_queue_max_size;			//used for statistics
+	unsigned int m_iterations;			//used for statistics
 
 	SortedSources m_sources;
 };
